requirements.txt
Django==2.0.5
django-jet==1.0.7
django-model-utils==3.1.2
django-widget-tweaks==1.4.2
easy-thumbnails==2.5
Pillow==5.1.0
psycopg2-binary==2.7.4
pytz==2018.4
